+++ src/components/@index.js
@@ -1,7 +1,7 @@
 import AlignContainer from './align_container';
 import Button from './button';
 import { InfiniteList, PostCard, Preview } from './infinite_list/@index';
-import Layout from './layout';
+import { Layout } from './layout/@index';
 import Poster from './poster';
 import Skeleton from './skeleton';
 import TopButton from './top_button';
