+++ src/components/layout/@index.js
@@ -1,0 +1,2 @@
+import Layout from './layout';
+export { Layout };
