+++ src/libs/react_router_dom/router.jsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from 'react-router-dom';
-import Layout from '../../components/layout';
+import Layout from '../../components/layout/layout';
 import { BoardPage, DetailPage, HomePage } from '../../pages/@index';
 
 const router = createBrowserRouter([
